requirements.txt
# This file is @generated by PDM.
# Please do not edit it manually.

alembic==1.16.4
bandit==1.9.4
beautifulsoup4==4.13.4
blinker==1.9.0
boolean-py==5.0
cachecontrol[filecache]==0.14.3
certifi==2025.7.9
cfgv==3.5.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6; platform_system == "Windows"
cyclonedx-python-lib==9.1.0
defusedxml==0.7.1
distlib==0.4.0
docker==7.1.0
filelock==3.25.1
flask==3.1.3
flask-migrate==4.1.0
flask-sqlalchemy==3.1.1
flask-talisman==1.1.0
flask-wtf==1.2.2
greenlet==3.2.3
html-sanitizer==2.6.0
identify==2.6.15
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.6
license-expression==30.4.3
lxml==6.0.1
lxml-html-clean==0.4.2
mako==1.3.10
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
mistune==3.1.4
msgpack==1.1.1
nodeenv==1.10.0
packageurl-python==0.17.1
packaging==25.0
pbr==6.1.1
peewee==3.18.2
pip==26.0.1
pip-api==0.0.34
pip-audit==2.10.0
pip-requirements-parser==32.0.1
platformdirs==4.3.8
pre-commit==4.5.1
py-serializable==2.0.0
pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pywin32==311; sys_platform == "win32"
pyyaml==6.0.2
requests==2.32.4
rich==14.0.0
ruff==0.15.5
scout==3.0.3
setuptools==80.9.0
six==1.17.0
sortedcontainers==2.4.0
soupsieve==2.7
sqlalchemy==2.0.41
stevedore==5.4.1
toml==0.10.2
tomli==2.4.0
tomli-w==1.2.0
typing-extensions==4.14.1
urllib3==2.6.3
virtualenv==20.36.1
werkzeug==3.1.6
wheel==0.46.3
whoosh==2.7.4
wtforms==3.2.1
